backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
pytest
psycopg[binary]
psycopg-pool
python-dotenv
